+++ fixtures/file-full.yaml
@@ -1,6 +1,6 @@
+full: true
 file:
   - type: $.InstanceType
     id: $.InstanceId
-    full: true
     paths:
       - fixtures/config_change.json
